More compile breakage. Review URL: https://chromiumcodereview.appspot.com/9303030
git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@10565 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/spaces.cc
@@ -283,7 +283,8 @@ bool MemoryAllocator::SetUp(intptr_t capacity, intptr_t capacity_executable) {
 
 void MemoryAllocator::TearDown() {
   // Check that spaces were torn down before MemoryAllocator.
-  CHECK_EQ(memory_allocator_reserved_, static_cast<size_t>(0));
+  CHECK_EQ(static_cast<int64_t>(memory_allocator_reserved_),
+           static_cast<int64_t>(0));
   // TODO(gc) this will be true again when we fix FreeMemory.
   // ASSERT(size_executable_ == 0);
   capacity_ = 0;
